rename package pkg/stream to uplink/stream
