+++ lib/uplink/bucket.go
@@ -13,7 +13,7 @@ import (
 	"storj.io/storj/pkg/metainfo/kvmetainfo"
 	"storj.io/storj/pkg/storage/streams"
 	"storj.io/storj/pkg/storj"
-	"storj.io/storj/pkg/stream"
+	"storj.io/storj/uplink/stream"
 )
 
 // Bucket represents operations you can perform on a bucket
