+++ lib/uplink/object.go
@@ -12,7 +12,7 @@ import (
 	"storj.io/storj/pkg/metainfo/kvmetainfo"
 	"storj.io/storj/pkg/storage/streams"
 	"storj.io/storj/pkg/storj"
-	"storj.io/storj/pkg/stream"
+	"storj.io/storj/uplink/stream"
 )
 
 // ObjectMeta contains metadata about a specific Object.
