+++ pkg/metainfo/kvmetainfo/objects_test.go
@@ -18,7 +18,7 @@ import (
 	"storj.io/storj/pkg/metainfo/kvmetainfo"
 	"storj.io/storj/pkg/storage/streams"
 	"storj.io/storj/pkg/storj"
-	"storj.io/storj/pkg/stream"
+	"storj.io/storj/uplink/stream"
 )
 
 const TestFile = "test-file"
