+++ uplink/stream/common.go
